+++ src/main/java/com/cts/FoodChainX/repository/DeliveryRepository.java
@@ -6,20 +6,20 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.cts.FoodChainX.model.Delivery;
 
-public interface DeliveryRepository extends JpaRepository<Delivery, Integer> {
+public interface DeliveryRepository extends JpaRepository<Delivery, Long> {
 
     
-    List<Delivery> findByShipmentId(Integer shipmentId);
+    List<Delivery> findByShipmentId(Long shipmentId);
 
     
-    List<Delivery> findByRetailerId(Integer retailerId);
+    List<Delivery> findByRetailerId(Long retailerId);
 
     
     List<Delivery> findByStatus(String status);
 
     
-    List<Delivery> findByRetailerIdOrderByDateDesc(Integer retailerId);
+    List<Delivery> findByRetailerIdOrderByDateDesc(Long retailerId);
 
     
-    List<Delivery> findByRetailerIdAndStatus(Integer retailerId, String status);
+    List<Delivery> findByRetailerIdAndStatus(Long retailerId, String status);
 }
